@@ -1,53 +1,60 @@
 {
-  "name": "seatunnel-ui",
+  "name": "seatunnel-engine-ui",
   "version": "0.0.0",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build --mode production",
+    "build": "run-p type-check \"build-only {@}\" --",
     "preview": "vite preview",
-    "lint": "eslint src --fix --ext .ts,.tsx,.vue",
-    "prettier": "prettier --write \"src/**/*.{vue,ts,tsx}\""
+    "test:unit": "vitest",
+    "test:e2e": "start-server-and-test preview http://localhost:4173 'cypress run --e2e'",
+    "test:e2e:dev": "start-server-and-test 'vite dev --port 4173' http://localhost:4173 'cypress open --e2e'",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "@antv/layout": "0.1.31",
-    "@antv/x6": "1.30.1",
-    "@antv/x6-vue-shape": "1.5.3",
-    "@vueuse/core": "^9.13.0",
-    "autoprefixer": "^10.4.13",
-    "axios": "^1.3.4",
-    "date-fns": "^2.29.3",
-    "date-fns-tz": "^2.0.0",
-    "naive-ui": "2.34.3",
+    "autoprefixer": "^10.4.20",
+    "axios": "^1.7.7",
+    "date-fns": "^3.6.0",
+    "date-fns-tz": "^3.1.3",
+    "naive-ui": "^2.39.0",
     "nprogress": "^0.2.0",
-    "pinia": "^2.0.32",
-    "pinia-plugin-persistedstate": "^3.1.0",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.7",
-    "vue": "^3.2.47",
-    "vue-i18n": "^9.2.2",
-    "vue-router": "^4.1.6"
+    "pinia": "^2.1.7",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.11",
+    "vue": "^3.4.29",
+    "vue-i18n": "^10.0.1",
+    "vue-router": "^4.3.3"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.191",
-    "@types/node": "^18.14.6",
-    "@types/nprogress": "^0.2.0",
-    "@typescript-eslint/eslint-plugin": "^5.54.0",
-    "@typescript-eslint/parser": "^5.54.0",
-    "@vicons/antd": "^0.12.0",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "@vitejs/plugin-vue-jsx": "^3.0.0",
-    "dart-sass": "^1.25.0",
-    "eslint": "^8.35.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.9.0",
-    "prettier": "^2.8.4",
-    "sass": "^1.58.3",
-    "sass-loader": "^13.2.0",
-    "typescript": "^4.9.5",
-    "typescript-plugin-css-modules": "^4.2.2",
-    "vite": "^4.1.4",
-    "vite-plugin-compression": "^0.5.1",
-    "vue-tsc": "^1.2.0"
+    "@pinia/testing": "^0.1.5",
+    "@rushstack/eslint-patch": "^1.8.0",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^20.14.5",
+    "@types/nprogress": "^0.2.3",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vitejs/plugin-vue-jsx": "^4.0.0",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/test-utils": "^2.4.6",
+    "@vue/tsconfig": "^0.5.1",
+    "cypress": "^13.12.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-cypress": "^3.3.0",
+    "eslint-plugin-vue": "^9.23.0",
+    "jsdom": "^24.1.0",
+    "npm-run-all2": "^6.2.0",
+    "prettier": "^3.2.5",
+    "sass-embedded": "^1.78.0",
+    "start-server-and-test": "^2.0.4",
+    "typescript": "~5.4.0",
+    "vite": "^5.3.1",
+    "vite-plugin-vue-devtools": "^7.3.1",
+    "vitest": "^1.5.3",
+    "vue-tsc": "^2.0.21"
   }
 }
