@@ -21,9 +21,7 @@ import NProgress from 'nprogress'
 import 'nprogress/nprogress.css'
 
 const router = createRouter({
-  history: createWebHistory(
-    import.meta.env.MODE === 'production' ? '/ui/' : '/'
-  ),
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes
 })
 
