@@ -16,49 +16,49 @@
  */
 
 export interface Monitor {
-  "processors": string;
-  "physical.memory.total": string;
-  "physical.memory.free": string;
-  "swap.space.total": string;
-  "swap.space.free": string;
-  "heap.memory.used": string;
-  "heap.memory.free": string;
-  "heap.memory.total": string;
-  "heap.memory.max": string;
-  "heap.memory.used/total": string;
-  "heap.memory.used/max": string;
-  "minor.gc.count": string;
-  "minor.gc.time": string;
-  "major.gc.count": string;
-  "major.gc.time": string;
-  "load.process": string;
-  "load.system": string;
-  "load.systemAverage": string;
-  "thread.count": string;
-  "thread.peakCount": string;
-  "cluster.timeDiff": string;
-  "event.q.size": string;
-  "executor.q.async.size": string;
-  "executor.q.client.size": string;
-  "executor.q.client.query.size": string;
-  "executor.q.client.blocking.size": string;
-  "executor.q.query.size": string;
-  "executor.q.scheduled.size": string;
-  "executor.q.io.size": string;
-  "executor.q.system.size": string;
-  "executor.q.operations.size": string;
-  "executor.q.priorityOperation.size": string;
-  "operations.completed.count": string;
-  "executor.q.mapLoad.size": string;
-  "executor.q.mapLoadAllKeys.size": string;
-  "executor.q.cluster.size": string;
-  "executor.q.response.size": string;
-  "operations.running.count": string;
-  "operations.pending.invocations.percentage": string;
-  "operations.pending.invocations.count": string;
-  "proxy.count": string;
-  "clientEndpoint.count": string;
-  "connection.active.count": string;
-  "client.connection.count": string;
-  "connection.count": string;
+  processors: string
+  'physical.memory.total': string
+  'physical.memory.free': string
+  'swap.space.total': string
+  'swap.space.free': string
+  'heap.memory.used': string
+  'heap.memory.free': string
+  'heap.memory.total': string
+  'heap.memory.max': string
+  'heap.memory.used/total': string
+  'heap.memory.used/max': string
+  'minor.gc.count': string
+  'minor.gc.time': string
+  'major.gc.count': string
+  'major.gc.time': string
+  'load.process': string
+  'load.system': string
+  'load.systemAverage': string
+  'thread.count': string
+  'thread.peakCount': string
+  'cluster.timeDiff': string
+  'event.q.size': string
+  'executor.q.async.size': string
+  'executor.q.client.size': string
+  'executor.q.client.query.size': string
+  'executor.q.client.blocking.size': string
+  'executor.q.query.size': string
+  'executor.q.scheduled.size': string
+  'executor.q.io.size': string
+  'executor.q.system.size': string
+  'executor.q.operations.size': string
+  'executor.q.priorityOperation.size': string
+  'operations.completed.count': string
+  'executor.q.mapLoad.size': string
+  'executor.q.mapLoadAllKeys.size': string
+  'executor.q.cluster.size': string
+  'executor.q.response.size': string
+  'operations.running.count': string
+  'operations.pending.invocations.percentage': string
+  'operations.pending.invocations.count': string
+  'proxy.count': string
+  'clientEndpoint.count': string
+  'connection.active.count': string
+  'client.connection.count': string
+  'connection.count': string
 }
