@@ -15,14 +15,12 @@
  * limitations under the License.
  */
 
-type Locales = 'zh_CN' | 'en_US'
+export type Locales = 'zh_CN' | 'en_US'
 
-interface SettingStore {
+export interface SettingStore {
   sequenceColumn: boolean
   dataUniqueValue: boolean
   fillet: number
   requestTime: number
   locales: Locales
 }
-
-export { SettingStore, Locales }
