@@ -17,14 +17,14 @@
 
 import i18n from '@/locales'
 const { t } = i18n.global
-export const getRemainTime = (milliseconds: number): any => {
-    if (!milliseconds) return
+export const getRemainTime = (milliseconds: number): string => {
+    if (!milliseconds) return ''
     milliseconds = milliseconds / 1000
     let d = parseInt(milliseconds / 60 / 60 / 24 + '')
     let h = parseInt(milliseconds / 60 / 60 % 24 + '')
     let m = parseInt(milliseconds / 60 % 60 + '')
     let s = parseInt(milliseconds % 60 + '')
-    const dText = h > 0 ? `${d}${t('common.date')} ` : ''
+    const dText = d > 0 ? `${d}${t('common.date')} ` : ''
     const hText = h > 0 ? `${h}${t('common.hour')} ` : ''
     const mText = m > 0 ? `${m}${t('common.min')} ` : ''
     const sText = s > 0 ? `${s}${t('common.second')}` : `0${t('common.second')}`
