@@ -16,65 +16,72 @@
  */
 
 import { defineComponent, getCurrentInstance, h, ref } from 'vue'
-import { useMessage, NDataTable } from 'naive-ui'
+import { useMessage, NDataTable, NTag } from 'naive-ui'
 import { useI18n } from 'vue-i18n'
-import { getFinishedJobs } from '@/service/job'
+import { JobsService } from '@/service/job'
 import type { DataTableColumns } from 'naive-ui'
 import { NButton } from 'naive-ui'
-import { Job } from '@/service/job/types'
+import type { Job } from '@/service/job/types'
 
 export default defineComponent({
-    setup() {
-        const { t } = useI18n()
+  setup() {
+    const { t } = useI18n()
 
-        const jobs = ref([] as Job[])
+    const jobs = ref([] as Job[])
 
-        getFinishedJobs().then(res => jobs.value = res)
-        function createColumns(): DataTableColumns<Job> {
-            const view = (job: Job) => { }
-            return [
-                {
-                    title: 'No',
-                    key: 'no'
-                },
-                {
-                    title: 'Title',
-                    key: 'title'
-                },
-                {
-                    title: 'Length',
-                    key: 'length'
-                },
-                {
-                    title: 'Action',
-                    key: 'actions',
-                    render(row) {
-                        return h(
-                            NButton,
-                            {
-                                strong: true,
-                                tertiary: true,
-                                size: 'small',
-                                onClick: () => view(row)
-                            },
-                            { default: () => 'View' }
-                        )
-                    }
-                }
-            ]
+    JobsService.getFinishedJobs().then((res) => (jobs.value = res))
+    function createColumns(): DataTableColumns<Job> {
+      const view = (job: Job) => {}
+      return [
+        {
+          title: 'No',
+          key: 'No',
+          render: (row, index) => h('div', index + 1)
+        },
+        {
+          title: 'Id',
+          key: 'jobId'
+        },
+        {
+          title: 'Name',
+          key: 'jobName'
+        },
+        {
+          title: 'Create Time',
+          key: 'createTime'
+        },
+        {
+          title: 'Status',
+          key: 'jobStatus',
+          render(row) {
+            return h(NTag, { bordered: false, type: 'success' }, { default: () => row.jobStatus })
+          }
+        },
+        {
+          title: 'Action',
+          key: 'actions',
+          render(row) {
+            return h(
+              NButton,
+              {
+                strong: true,
+                tertiary: true,
+                size: 'small',
+                onClick: () => view(row)
+              },
+              { default: () => 'View' }
+            )
+          }
         }
-
-        const columns = createColumns()
-        return () => (
-            <div class='w-full bg-white px-12 pt-6 pb-12 border border-gray-100 rounded-xl'>
-                <h2 class='font-bold text-2xl pb-6'>{t('jobs.finishedJobs')}</h2>
-                <NDataTable
-                    columns={columns}
-                    data={jobs.value}
-                    pagination={false}
-                    bordered={false}
-                />
-            </div>
-        )
+      ]
     }
+
+    const columns = createColumns()
+    return () => (
+      <div class="w-full bg-white px-12 pt-6 pb-12 border border-gray-100 rounded-xl">
+        <h2 class="font-bold text-2xl pb-6">{t('jobs.finishedJobs')}</h2>
+        <NDataTable columns={columns} data={jobs.value} pagination={false} bordered={false} />
+      </div>
+    )
+  }
 })
