@@ -21,69 +21,68 @@ import type { DataTableColumns } from 'naive-ui'
 import { NButton } from 'naive-ui'
 import { NSpace, NLayout, NLayoutContent } from 'naive-ui'
 import { getMonitors } from '@/service/manager'
-import { Monitor } from '@/service/manager/types'
+import type { Monitor } from '@/service/manager/types'
 
 export default defineComponent({
-    setup() {
+  setup() {
+    const { t } = useI18n()
 
-        const { t } = useI18n()
+    const monitors = ref([] as Monitor[])
 
-        const monitors = ref([] as Monitor[])
-
-        getMonitors().then(res => monitors.value = res)
-        function createColumns(): DataTableColumns<Monitor> {
-            const view = (job: Monitor) => { }
-            return [
-                {
-                    title: 'Host',
-                    key: 'host'
-                },
-                {
-                    title: 'Port',
-                    key: 'port'
-                },
-                {
-                    title: 'Physical MEM',
-                    key: 'physical.memory.total'
-                },
-                {
-                    title: 'Heap MEM Used',
-                    key: 'heap.memory.used'
-                },
-                {
-                    title: 'Action',
-                    key: 'actions',
-                    render(row) {
-                        return h(
-                            NButton,
-                            {
-                                strong: true,
-                                tertiary: true,
-                                size: 'small',
-                                onClick: () => view(row)
-                            },
-                            { default: () => 'View' }
-                        )
-                    }
-                }
-            ]
+    getMonitors().then((res) => (monitors.value = res))
+    function createColumns(): DataTableColumns<Monitor> {
+      const view = (job: Monitor) => {}
+      return [
+        {
+          title: 'Host',
+          key: 'host'
+        },
+        {
+          title: 'Port',
+          key: 'port'
+        },
+        {
+          title: 'Physical MEM',
+          key: 'physical.memory.total'
+        },
+        {
+          title: 'Heap MEM Used',
+          key: 'heap.memory.used'
+        },
+        {
+          title: 'Action',
+          key: 'actions',
+          render(row) {
+            return h(
+              NButton,
+              {
+                strong: true,
+                tertiary: true,
+                size: 'small',
+                onClick: () => view(row)
+              },
+              { default: () => 'View' }
+            )
+          }
         }
-
-        const columns = createColumns()
-        return () => (
-            <NLayout>
-                <NLayoutContent>
-                    <div class='w-full bg-white px-12 pt-6 pb-12 border border-gray-100 rounded-xl'>
-                        <h2 class='font-bold text-2xl pb-6'>{t('managers.managers')}</h2>
-                        <NDataTable
-                            columns={columns}
-                            data={monitors.value}
-                            pagination={false}
-                            bordered={false}
-                        />
-                    </div>
-                </NLayoutContent>
-            </NLayout>
-        )
+      ]
     }
+
+    const columns = createColumns()
+    return () => (
+      <NLayout>
+        <NLayoutContent>
+          <div class="w-full bg-white px-12 pt-6 pb-12 border border-gray-100 rounded-xl">
+            <h2 class="font-bold text-2xl pb-6">{t('managers.managers')}</h2>
+            <NDataTable
+              columns={columns}
+              data={monitors.value}
+              pagination={false}
+              bordered={false}
+            />
+          </div>
+        </NLayoutContent>
+      </NLayout>
+    )
+  }
 })
