@@ -15,37 +15,25 @@
  * limitations under the License.
  */
 
+import { fileURLToPath, URL } from 'node:url'
 import { defineConfig, loadEnv } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import vueJsx from '@vitejs/plugin-vue-jsx'
-import viteCompression from 'vite-plugin-compression'
-import path from 'path'
+import vueDevTools from 'vite-plugin-vue-devtools'
 
+// https://vitejs.dev/config/
 export default defineConfig({
-  base: process.env.NODE_ENV === 'production' ? '/ui/' : '/',
-  plugins: [
-    vue(),
-    vueJsx(),
-    viteCompression({
-      verbose: true,
-      disable: false,
-      threshold: 10240,
-      algorithm: 'gzip',
-      ext: '.gz',
-      deleteOriginFile: false
-    })
-  ],
+  base: '/ui/',
+  plugins: [vue(), vueJsx(), vueDevTools()],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, 'src'),
-      'vue-i18n': 'vue-i18n/dist/vue-i18n.cjs.js',
-      '@intlify/shared': '@intlify/shared/dist/shared.cjs.js'
+      '@': fileURLToPath(new URL('./src', import.meta.url))
     }
   },
   server: {
     proxy: {
       '/hazelcast/rest': {
-        target: loadEnv('development', './').VITE_APP_DEV_WEB_URL,
+        target: loadEnv('development', './').VITE_APP_API_SERVICE,
         changeOrigin: true
       }
     }
